refactor: move legacy state helpers to app
