+++ src/app.rs
@@ -328,6 +328,107 @@ impl DelegateEntry {
     }
 }
 
+/// Update per-delegation stats from a single event arriving on a child session.
+pub(crate) fn accumulate_delegate_stats(stats: &mut DelegateStats, kind: &EventKind) {
+    match kind {
+        EventKind::ToolCallStart { .. } => {
+            stats.tool_calls = stats.tool_calls.saturating_add(1);
+        }
+        EventKind::AssistantMessageStored { .. } => {
+            stats.messages = stats.messages.saturating_add(1);
+        }
+        EventKind::LlmRequestEnd {
+            cost_usd,
+            context_tokens,
+            ..
+        } => {
+            if let Some(c) = cost_usd {
+                stats.cost_usd += c;
+            }
+            if let Some(ctx) = context_tokens {
+                stats.context_tokens = *ctx;
+            }
+        }
+        EventKind::ProviderChanged {
+            context_limit: Some(limit),
+            ..
+        } => {
+            stats.context_limit = *limit;
+        }
+        EventKind::LlmRequestStart { .. }
+        | EventKind::SnapshotStart { .. }
+        | EventKind::SnapshotEnd { .. }
+        | EventKind::ProgressRecorded { .. }
+        | EventKind::ArtifactRecorded { .. }
+        | EventKind::SessionQueued { .. }
+        | EventKind::SessionConfigured { .. }
+        | EventKind::ToolsAvailable { .. }
+        | EventKind::SessionCreated
+        | EventKind::Unknown => {}
+        _ => {}
+    }
+}
+
+pub(crate) fn update_delegate_child_state(state: &mut DelegateChildState, kind: &EventKind) {
+    match kind {
+        EventKind::ElicitationRequested {
+            elicitation_id,
+            message,
+            source,
+            requested_schema,
+            ..
+        } => {
+            *state = DelegateChildState::PendingElicitation {
+                elicitation_id: elicitation_id.clone(),
+                message: message.clone(),
+                requested_schema: requested_schema.clone(),
+                source: source.clone(),
+            };
+        }
+        EventKind::ToolCallEnd { tool_name, .. } if tool_name == "question" => {
+            *state = DelegateChildState::QuestionToolFinished;
+        }
+        EventKind::AssistantMessageStored { .. } => {
+            *state = DelegateChildState::AssistantMessage;
+        }
+        EventKind::UserMessageStored { .. } => {
+            *state = DelegateChildState::UserMessage;
+        }
+        EventKind::ToolCallStart { .. }
+        | EventKind::AssistantContentDelta { .. }
+        | EventKind::AssistantThinkingDelta { .. }
+        | EventKind::PromptReceived { .. }
+        | EventKind::LlmRequestStart { .. }
+        | EventKind::LlmRequestEnd { .. }
+        | EventKind::CompactionStart { .. }
+        | EventKind::CompactionEnd { .. }
+        | EventKind::SnapshotStart { .. }
+        | EventKind::SnapshotEnd { .. }
+        | EventKind::ProgressRecorded { .. }
+        | EventKind::ArtifactRecorded { .. }
+        | EventKind::SessionQueued { .. }
+        | EventKind::ProviderChanged { .. }
+        | EventKind::Error { .. }
+        | EventKind::Cancelled => {
+            *state = DelegateChildState::OtherProgress;
+        }
+        EventKind::TurnStarted
+        | EventKind::SessionModeChanged { .. }
+        | EventKind::SessionConfigured { .. }
+        | EventKind::ToolsAvailable { .. }
+        | EventKind::SessionCreated
+        | EventKind::DelegationRequested { .. }
+        | EventKind::DelegationCompleted { .. }
+        | EventKind::DelegationFailed { .. }
+        | EventKind::DelegationCancelled { .. }
+        | EventKind::SessionForked { .. }
+        | EventKind::Unknown => {}
+        EventKind::ToolCallEnd { .. } => {
+            *state = DelegateChildState::OtherProgress;
+        }
+    }
+}
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DelegateStatus {
     InProgress,
@@ -381,6 +482,39 @@ pub enum ChatEntry {
         /// None = pending; Some = responded with this outcome label.
         outcome: Option<String>,
     },
+}
+
+pub(crate) fn backfill_elicitation_outcomes(messages: &mut [ChatEntry], result_str: &str) {
+    let Ok(val) = serde_json::from_str::<serde_json::Value>(result_str) else {
+        return;
+    };
+    let Some(answers) = val.get("answers").and_then(|a| a.as_array()) else {
+        return;
+    };
+
+    let mut answer_iter = answers.iter();
+    for entry in messages.iter_mut() {
+        let ChatEntry::Elicitation { outcome, .. } = entry else {
+            continue;
+        };
+        if outcome.as_deref() != Some("responded") {
+            continue;
+        }
+        let Some(answer_entry) = answer_iter.next() else {
+            break;
+        };
+        let labels: Vec<String> = answer_entry
+            .get("answers")
+            .and_then(|a| a.as_array())
+            .map(|arr| {
+                arr.iter()
+                    .filter_map(|v| v.as_str())
+                    .map(|s| format!("{OUTCOME_BULLET}{s}"))
+                    .collect()
+            })
+            .unwrap_or_default();
+        *outcome = Some(labels.join("\n"));
+    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -3325,7 +3459,6 @@ mod reasoning_effort_tests {
 #[cfg(test)]
 mod delegate_entry_tests {
     use super::*;
-    use crate::server_msg::accumulate_delegate_stats;
 
     fn make_entry(delegation_id: &str, objective: &str, status: DelegateStatus) -> DelegateEntry {
         DelegateEntry {
@@ -5741,7 +5874,6 @@ mod session_mode_tests {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::server_msg::backfill_elicitation_outcomes;
 
     fn make_turn(message_id: &str) -> UndoableTurn {
         UndoableTurn {
